@@ -15,29 +15,28 @@
   ],
   "scripts": {
     "clean": "ts-node tests/pullImageKillOld.ts",
-    "test": "yarn clean && jest tests/recovery.test.ts tests/standard.test.ts",
-    "testSR25519": "yarn clean && jest tests/sr25519.test.ts"
+    "test": "yarn clean && jest --maxConcurrency 3"
   },
   "dependencies": {
-    "@zondax/ledger-substrate": "^0.39.0",
-    "@zondax/zemu": "^0.34.0"
+    "@zondax/ledger-substrate": "^0.40.6",
+    "@zondax/zemu": "^0.42.1"
   },
   "devDependencies": {
-    "@types/jest": "^29.2.1",
+    "@types/jest": "^29.5.0",
     "@types/ledgerhq__hw-transport": "^4.21.4",
-    "@typescript-eslint/eslint-plugin": "^5.42.0",
-    "@typescript-eslint/parser": "^5.42.0",
+    "@typescript-eslint/eslint-plugin": "^5.56.0",
+    "@typescript-eslint/parser": "^5.56.0",
     "blakejs": "^1.1.1",
-    "crypto-js": "4.1.1",
+    "crypto-js": "^4.1.1",
     "ed25519-supercop": "^2.0.1",
-    "eslint": "^8.26.0",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint": "^8.36.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-jest": "^27.1.3",
     "eslint-plugin-prettier": "^4.0.0",
-    "jest": "29.2.2",
+    "jest": "^29.5.0",
     "jssha": "^3.2.0",
-    "prettier": "^2.5.1",
+    "prettier": "^2.8.6",
     "ts-jest": "^29.0.3",
     "ts-node": "^10.9.1",
     "typescript": "^4.5.3"
