@@ -18,9 +18,12 @@
 #include "base58.h"
 #include "coin.h"
 #include "cx.h"
-#include "rslib.h"
 #include "zxmacros.h"
 #include "ristretto.h"
+
+#ifdef SUPPORT_SR25519
+#include "rslib.h"
+#endif
 
 uint32_t hdPath[HDPATH_LEN_DEFAULT];
 
@@ -161,8 +164,13 @@ void zeroize_sr25519_signdata(void) {
     explicit_bzero(sr25519_signature, sizeof(sr25519_signature));
 }
 
-void copy_sr25519_signdata(uint8_t *buffer) {
+zxerr_t copy_sr25519_signdata(uint8_t *buffer, uint16_t bufferLen) {
+    if (SIG_PLUS_TYPE_LEN > bufferLen) {
+        return zxerr_buffer_too_small;
+    }
+
     memcpy(buffer, sr25519_signature, SIG_PLUS_TYPE_LEN);
+    return zxerr_ok;
 }
 
 static zxerr_t crypto_sign_sr25519_helper(const uint8_t *data, size_t len) {
