@@ -51,7 +51,7 @@ typedef enum {
 // Coin Specific
 #define PK_ADDRESS_TYPE COIN_ADDR_TYPE
 #define SUPPORTED_TX_VERSION_CURRENT LEDGER_MAJOR_VERSION
-#define SUPPORTED_TX_VERSION_PREVIOUS (LEDGER_MAJOR_VERSION - 1)
+#define SUPPORTED_TX_VERSION_PREVIOUS 20
 #define SUPPORTED_SPEC_VERSION (LEDGER_MINOR_VERSION + 0)
 #define SUPPORTED_MINIMUM_SPEC_VERSION 9300
 
